seperate scene from the visual system and refactor entity classes to their own file
