+++ src/systems/visual/mode.ts
@@ -1,5 +1,5 @@
 import { GraphVisualSystem } from './visual';
-import { EdgeEntity, NodeEntity } from './scene_manger'
+import { EdgeEntity, NodeEntity } from './entity'
 
 export interface Mode {
   begin(system: GraphVisualSystem): void;
@@ -13,7 +13,7 @@ export class NodeMode implements Mode {
   }
 
   end(system: GraphVisualSystem) {
-    let nodes = system.nodes;
+    let nodes = system.scene.nodes;
     nodes.forEach((n) => n.events.off('click'));
     system.stage.off('click');
   }
@@ -25,20 +25,20 @@ export class NodeMode implements Mode {
       if (pointer != null) {
         let node = new NodeEntity(pointer.x, pointer.y);
         node.events.on('click', () => {
-          system.removeNode(node);
-          system.markNeedsDrawing();
+          system.scene.removeNode(node);
+          system.scene.markNeedsDrawing();
         });
-        system.addNode(node);
-        system.markNeedsDrawing();
+        system.scene.addNode(node);
+        system.scene.markNeedsDrawing();
       }
     });
   }
 
   private configureDestruction(system: GraphVisualSystem) {
-    let nodes = system.nodes
+    let nodes = system.scene.nodes
     nodes.forEach((n) => n.events.on('click', () => {
-      system.removeNode(n);
-      system.markNeedsDrawing();
+      system.scene.removeNode(n);
+      system.scene.markNeedsDrawing();
     }));
   }
 }
@@ -50,16 +50,16 @@ export class EdgeMode implements Mode {
   }
 
   end(system: GraphVisualSystem) {
-    let nodes = system.nodes;
+    let nodes = system.scene.nodes;
     nodes.forEach((n) => n.events.off('click'));
-    let edges = system.edges;
+    let edges = system.scene.edges;
     edges.forEach((e) => e.events.off('click'));
   }
 
   create = (system: GraphVisualSystem) => {
     let from: NodeEntity | null = null;
     let to: NodeEntity | null = null;
-    let nodes = system.nodes;
+    let nodes = system.scene.nodes;
     nodes.forEach((n) => n.events.on('click', () => {
       if (from == null) {
         from = n;
@@ -70,21 +70,21 @@ export class EdgeMode implements Mode {
       if (from != null && to != null) {
         let edge = new EdgeEntity(from, to);
         edge.events.on('click', () => {
-          system.removeEdge(edge);
-          system.markNeedsDrawing();
+          system.scene.removeEdge(edge);
+          system.scene.markNeedsDrawing();
         });
-        system.addEdge(edge);
-        system.markNeedsDrawing();
+        system.scene.addEdge(edge);
+        system.scene.markNeedsDrawing();
         from = to = null;
       }
     }));
   }
 
   destroy = (system: GraphVisualSystem) => {
-    let edges = system.edges;
+    let edges = system.scene.edges;
     edges.forEach((e) => e.events.on('click', () => {
-      system.removeEdge(e);
-      system.markNeedsDrawing();
+      system.scene.removeEdge(e);
+      system.scene.markNeedsDrawing();
     }));
   }
 }
